+++ examples/src/contracts/owned_token.rs
@@ -1,7 +1,7 @@
 use odra::prelude::*;
 use odra::{
     casper_types::U256,
-    module::{Module, ModuleWrapper},
+    module::{Module, SubModule},
     Address
 };
 use odra_modules::access::Ownable;
@@ -9,8 +9,8 @@ use odra_modules::erc20::Erc20;
 
 #[odra::module]
 pub struct OwnedToken {
-    ownable: ModuleWrapper<Ownable>,
-    erc20: ModuleWrapper<Erc20>
+    ownable: SubModule<Ownable>,
+    erc20: SubModule<Erc20>
 }
 
 #[odra::module]