+++ modules/src/erc1155_token.rs
@@ -8,7 +8,7 @@ use crate::erc1155::Erc1155;
 use odra::prelude::*;
 use odra::{
     casper_types::{bytesrepr::Bytes, U256},
-    module::{Module, ModuleWrapper},
+    module::{Module, SubModule},
     Address
 };
 
@@ -16,8 +16,8 @@ use odra::{
 /// It uses the [ERC1155](Erc1155Base) base implementation and the [Ownable] module.
 #[odra::module(events = [TransferBatch, TransferSingle])]
 pub struct Erc1155Token {
-    core: ModuleWrapper<Erc1155Base>,
-    ownable: ModuleWrapper<Ownable>
+    core: SubModule<Erc1155Base>,
+    ownable: SubModule<Ownable>
 }
 
 #[odra::module]