+++ odra-macros/src/ast/module_def.rs
@@ -42,9 +42,9 @@ mod tests {
         let def = ModuleDefItem::try_from(&ir).unwrap();
         let expected = quote::quote! {
             pub struct CounterPack {
-                counter0: ModuleWrapper<Counter>,
-                counter1: ModuleWrapper<Counter>,
-                counter2: ModuleWrapper<Counter>,
+                counter0: SubModule<Counter>,
+                counter1: SubModule<Counter>,
+                counter2: SubModule<Counter>,
                 counters: Variable<u32>,
                 counters_map: Mapping<u8, Counter>,
                 __env: odra::prelude::Rc<odra::ContractEnv>