+++ odra-macros/src/test_utils.rs
@@ -95,9 +95,9 @@ pub mod mock {
     pub fn module_definition() -> ModuleStructIR {
         let module = quote!(
             pub struct CounterPack {
-                counter0: ModuleWrapper<Counter>,
-                counter1: ModuleWrapper<Counter>,
-                counter2: ModuleWrapper<Counter>,
+                counter0: SubModule<Counter>,
+                counter1: SubModule<Counter>,
+                counter2: SubModule<Counter>,
                 counters: Variable<u32>,
                 counters_map: Mapping<u8, Counter>
             }